+++ packages/ramps-controller/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `status` field to `ResourceState` to distinguish between uninitialized and empty-fetched states ([#8116](https://github.com/MetaMask/core/pull/8116))
+- Add `status` field to `ResourceState` to distinguish between uninitialized and empty-fetched states ([#8158](https://github.com/MetaMask/core/pull/8158))
 
 ## [10.2.0]